+++ spec/document/elements/paragraph/properties/tabs_spec.rb
@@ -6,4 +6,9 @@ describe 'My behaviour' do
     expect(docx.elements.first.paragraph_properties.tabs[0].position).to eq(OoxmlParser::OoxmlSize.new(1440))
     expect(docx.elements.first.paragraph_properties.tabs[0].value).to eq(:left)
   end
+
+  it 'tab_leader' do
+    docx = OoxmlParser::Parser.parse('spec/document/elements/paragraph/properties/tabs/tab_leader.docx')
+    expect(docx.elements.first.paragraph_properties.tabs[0].leader).to eq(:dot)
+  end
 end